requirements.txt
asyncio-nats-client==0.11.5
coverage==7.2.7
exceptiongroup==1.1.1
grpcio==1.54.2
grpcio-tools==1.54.2
iniconfig==2.0.0
nats-python==0.8.0
packaging==23.1
pluggy==1.0.0
protobuf==4.23.2
py==1.11.0
pytest==7.3.2
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-html==3.2.0
pytest-metadata==3.0.0
tomli==2.0.1
